requirements.txt
paho-mqtt
psycopg2
pystray
pillow
pywin32
winshell